# web/openapi.json
{
  "components": {
    "securitySchemes": {
      "session": {
        "in": "cookie",
        "name": "isb_session",
        "type": "apiKey"
      },
      "token": {
        "description": "An API token (isb token create)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "isb",
    "version": "0.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tools/org_create": {
      "post": {
        "description": "Platform admins: create an org (an incus project with its own bridge and network ACL), with optional limits and egress exceptions. Fails if it exists. Bind roots are set from the host's CLI only.",
        "operationId": "org_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "cpus": {
                    "description": "CPUs across the org's instances.",
                    "minimum": 1,
                    "type": "integer"
                  },
                  "default_cpus": {
                    "description": "CPUs an instance gets when its spec sets none.",
                    "minimum": 1,
                    "type": "integer"
                  },
                  "default_memory": {
                    "description": "Memory an instance gets when its spec sets none, e.g. 512MiB.",
                    "type": "string"
                  },
                  "disk": {
                    "description": "Disk across the org, e.g. 100GiB.",
                    "type": "string"
                  },
                  "egress": {
                    "description": "Private destinations the org may reach, CIDR[:PORTS[/tcp|udp]] (docs/orgs.md). Replaces the list; [] clears it.",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "instances": {
                    "description": "Instances in the org.",
                    "minimum": 1,
                    "type": "integer"
                  },
                  "memory": {
                    "description": "Memory across the org, e.g. 16GiB.",
                    "type": "string"
                  },
                  "org": {
                    "description": "The new org's name: [a-z0-9-], starts with a letter.",
                    "type": "string"
                  },
                  "placement": {
                    "description": "Where the org runs, set at creation: \"local\" (this host: an incus project sharing its kernel), {\"server\": NAME} (another host, server_list), or {\"vm\": {\"cpus\", \"memory\", \"disk\"}} (a dedicated VM this control plane makes on its own host: the org's own kernel; defaults 2 CPUs, 4GiB, 40GiB). An org is not moved afterwards.",
                    "oneOf": [
                      {
                        "enum": [
                          "local"
                        ],
                        "type": "string"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "server": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "server"
                        ],
                        "type": "object"
                      },
                      {
                        "additionalProperties": false,
                        "properties": {
                          "vm": {
                            "additionalProperties": false,
                            "properties": {
                              "cpus": {
                                "maximum": 256,
                                "minimum": 1,
                                "type": "integer"
                              },
                              "disk": {
                                "description": "At least 10GiB (default 40GiB).",
                                "type": "string"
                              },
                              "memory": {
                                "description": "At least 2GiB (default 4GiB).",
                                "type": "string"
                              }
                            },
                            "type": "object"
                          }
                        },
                        "required": [
                          "vm"
                        ],
                        "type": "object"
                      }
                    ]
                  },
                  "server": {
                    "description": "Where the org runs: local (default) or a server's name (server_list). Set at creation; an org is not moved between servers. Same as placement {\"server\": NAME}.",
                    "type": "string"
                  },
                  "wait": {
                    "description": "With a dedicated VM: wait until it is made and the org created (default true; minutes). false answers at once with `provision`; follow it with server_provision_get (name vm-<org>).",
                    "type": "boolean"
                  }
                },
                "required": [
                  "org"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "result": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "The tool's result"
          },
          "default": {
            "description": "An error: {error, message, data}"
          }
        },
        "summary": "Create an org"
      }
    },
...
  },
  "security": [
    {
      "token": []
    },
    {
      "session": []
    }
  ]
}